pyproject.toml
[project]
name = "graduation-microservices"
version = "0.1.0"
dependencies = [
    "cachetools==7.1.1",
    "faiss-cpu==1.13.2",
    "fastapi==0.129.1",
    "json-repair==0.44.1",
    "langchain==1.2.10",
    "langchain-cohere==0.5.1",
    "langchain-community==0.4.1",
    "langchain-experimental==0.4.1",
    "langchain-google-genai==4.2.2",
    "langchain-groq==1.1.2",
    "langchain-huggingface==1.2.0",
    "langchain-mistralai==1.1.4",
    "langchain-neo4j==0.8.0",
    "langchain-openai==1.2.1",
    "langgraph==1.0.9",
    "neo4j==6.1.0",
    "pandas>=2.0.0",
    "pydantic-settings==2.14.0",
    "pytest==9.0.3",
    "python-dotenv==1.2.1",
    "rank-bm25==0.2.2",
    "sentence-transformers==5.4.1",
    "uvicorn==0.41.0",
]

[tool.uv.workspace]
members = [
    "gateway",
    "orchestrator",
    "llm-service",
    "retrieval-service",
    "kg-service",
    "ingestion-service",
    "shared",
]
